pyproject.toml
[project]
name = "hoteles-estelar-chat"
version = "0.2.0"
description = "Chatbot corporativo para Hoteles Estelar S.A. con Q&A, resumen ejecutivo y generación de FAQ basados en reportes financieros."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-anthropic>=1.4.2",
    "langchain-core>=1.3.2",
    "langchain-ollama>=1.1.0",
    "langgraph>=0.2.28",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "streamlit>=1.57.0",
]

[project.optional-dependencies]
scraping = [
    "beautifulsoup4>=4.14.3",
    "playwright>=1.53.0",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
    "D",    # pydocstyle (docstrings)
]
ignore = [
    "D100",  # Missing docstring in module (handled separately)
    "D104",  # Missing docstring in package
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["llm", "app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
